refactor: modularise num-conversions
