+++ number-conversions/convert-base.js
@@ -1,0 +1,96 @@
+/**
+ * Helper function for the converting any base to any base
+ */
+function parseToDigitsArray(str, base) {
+  const digits = str.split('');
+  const ary = [];
+  for (let i = digits.length - 1; i >= 0; i -= 1) {
+    const n = parseInt(digits[i], base);
+    if (Number.isNaN(n)) return null;
+    ary.push(n);
+  }
+  return ary;
+}
+
+/**
+ * Helper function for the converting any base to any base
+ */
+function add(x, y, base) {
+  const z = [];
+  const n = Math.max(x.length, y.length);
+  let carry = 0;
+  let i = 0;
+  while (i < n || carry) {
+    const xi = i < x.length ? x[i] : 0;
+    const yi = i < y.length ? y[i] : 0;
+    const zi = carry + xi + yi;
+    z.push(zi % base);
+    carry = Math.floor(zi / base);
+    i += 1;
+  }
+  return z;
+}
+
+/**
+ * Helper function for the converting any base to any base
+ * Returns a*x, where x is an array of decimal digits and a is an ordinary
+ * JavaScript number. base is the number base of the array x.
+ */
+function multiplyByNumber(num, x, base) {
+  if (num < 0) return null;
+  if (num === 0) return [];
+
+  let result = [];
+  let power = x;
+
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    if (num & 1) {
+      result = add(result, power, base);
+    }
+
+    // eslint-disable-next-line no-param-reassign
+    num >>= 1;
+    if (num === 0) break;
+    power = add(power, power, base);
+  }
+  return result;
+}
+
+/**
+ * Helper function for the converting any base to any base
+ */
+function convertBase(str, fromBase, toBase) {
+  const digits = parseToDigitsArray(str, fromBase);
+  if (digits === null) return null;
+
+  let outArray = [];
+  let power = [1];
+  for (let i = 0; i < digits.length; i += 1) {
+    // invariant: at this point, fromBase^i = power
+    if (digits[i]) {
+      outArray = add(outArray, multiplyByNumber(digits[i], power, toBase), toBase);
+    }
+    power = multiplyByNumber(fromBase, power, toBase);
+  }
+
+  let out = '';
+  for (let i = outArray.length - 1; i >= 0; i -= 1) {
+    out += outArray[i].toString(toBase);
+  }
+  // if the original input was equivalent to zero, then 'out' will still be empty ''. Let's check for zero.
+  if (out === '') {
+    let sum = 0;
+    for (let i = 0; i < digits.length; i += 1) {
+      sum += digits[i];
+    }
+    if (sum === 0) out = '0';
+  }
+
+  return out;
+}
+
+module.exports = {
+  convertBase,
+  parseToDigitsArray,
+};
